refactor(edgeless): move  and  to std.ts
