+++ packages/blocks/src/page-block/edgeless/components/component-toolbar/change-shape-button.ts
@@ -7,11 +7,11 @@ import { DisposableGroup } from '@blocksuite/store';
 import { css, html, LitElement } from 'lit';
 import { customElement, property, query } from 'lit/decorators.js';
 
+import { countBy, maxBy } from '../../../../__internal__/utils/std.js';
 import type { ShapeMouseMode } from '../../../../__internal__/utils/types.js';
 import type { EdgelessShapeMenu } from '../../toolbar/shape-tool/shape-menu.js';
 import { ShapeComponentConfigMap } from '../../toolbar/shape-tool/shape-menu-config.js';
 import { createButtonPopper } from '../utils.js';
-import { countBy, maxBy } from './utils.js';
 
 function getMostCommonShape(
   elements: ShapeElement[]
